(WIP) 1.3 updates (june 21)
- Updated .env.example to add required lines
- Some logic for the AcknowledgementHandler
- Added AcknowledgeThis record to store the info for acknowledgements
- Change snapshotAgeOfServer type to long
- Some logic for the AnalyticsHandler
- Created AudioHandler for audio-related things
- Changed the Brew descriptions for the subcommands
- Updated BrewerServerDataHandler to configure for acknowledgements
- Added footer in the preview page
- Used new markdown styling for the preview page
- Verify the names of categories, channels, and roles won't get us yeeted from OpenAI
- Added more edge cases in the event the JSON is unable to be parsed for text before and after the brace closing
- Moved SUPPORTED_EXTENSIONS list to CustomOpenAiService
- Wrote logic for a CustomSettings command
- Updated AboutInformation to use the new domains
- Updated Launcher to reset the acknowledgements back to false after launching
- Added a parameter to isContentFlagged in OpenAIHandler to optionally opt out of flagging or not
- Minor edit to the rename prompt
- Added methods to get the remaining runs for a server/user
- Added buttons if using a transcription on VCs
- Some logic for VC transcriptions
- Indicate if a context transcription is cached
- Setter methods in UserSession changed to return itself for method chaining
- Moved VoiceHandler to commands/audio and migrated part of the code to its own class
- VoteHandler adds a 60 minute delay at the start and uses Instant.now().getEpochSecond() instead of System.currentTimeMillis() / 1000L
- Update openai-java to 0.13.0
- Added default JSONs for acknowledgements in the resources folder

+++ src/main/java/dev/alphaserpentis/bots/brewer/data/brewer/Analytics.java
@@ -5,10 +5,10 @@ import java.util.EnumMap;
 public class Analytics {
     private final EnumMap<ServiceType, Integer> usagePerServiceType = new EnumMap<>(ServiceType.class);
     private final int snapshotUserCount;
-    private final int snapshotAgeOfServer;
+    private final long snapshotAgeOfServer;
     private final PaidTier paidTier;
 
-    public Analytics(int snapshotUserCount, int snapshotAgeOfServer, PaidTier paidTier) {
+    public Analytics(int snapshotUserCount, long snapshotAgeOfServer, PaidTier paidTier) {
         this.snapshotUserCount = snapshotUserCount;
         this.snapshotAgeOfServer = snapshotAgeOfServer;
         this.paidTier = paidTier;
@@ -22,7 +22,7 @@ public class Analytics {
         return snapshotUserCount;
     }
 
-    public int getSnapshotAgeOfServer() {
+    public long getSnapshotAgeOfServer() {
         return snapshotAgeOfServer;
     }
